adapt test add_w to work with changed ratio calc
The changed ratio calculation changes the behaviour.
This commit modifies the add window test.
A sequence of images that draws the window layout that
will be visible on the screen with the changed ratio calculation,
compared to the layout that is visible without it, is uploaded on github.
I extracted the data for the images with a modified version of this
test run by writing the "layout_info" in a file
together with the expected "layout_info".
This test change is probably less worrying than the other one
which changes the resizing test.