adapt test resize to work with changed ratio calc
The changed ratio calculation changes the behaviour.
This commit modifies the "test_resizing" test.
A sequence of images that draws the window layout that
will be visible on the screen with the changed ratio calculation,
compared to the layout that is visible without it, is uploaded on github.
I extracted the data for the images with a modified version of this
test by writing the "layout_info" in a file
together with the expected "layout_info".

I made this test a bit more extensive than the original one.